Move `increased_trustees` fixture to `sovtoken/conftest`.
Moved `increased_trustees` from `sovtokenfees/conftest` to
`sovtoken/conftest`.

Fixed an error where the did/verkey didn't match the one on the
ledger for the new trustees.
