+++ sovtokenfees/sovtokenfees/test/conftest.py
@@ -2,9 +2,6 @@ from sovtoken.constants import XFER_PUBLIC, RESULT
 from sovtoken.main import integrate_plugin_in_node as enable_token
 from sovtokenfees.main import integrate_plugin_in_node as enable_fees
 
-from plenum.common.util import randomString
-from plenum.common.constants import TARGET_NYM, TRUSTEE_STRING, VERKEY
-
 # fixtures, do not remove
 from plenum.test.conftest import *
 from plenum import PLUGIN_CLIENT_REQUEST_FIELDS
@@ -72,36 +69,3 @@ def helpers(
 @pytest.fixture(autouse=True)
 def reset_fees(helpers):
     helpers.node.reset_fees()
-
-
-@pytest.fixture()
-def increased_trustees(helpers, trustee_wallets, sdk_wallet_trustee):
-    seeds = [randomString(32) for _ in range(3)]
-
-    requests = [
-        helpers.request.nym(seed=seed, role=TRUSTEE_STRING)
-        for seed in seeds
-    ]
-
-    responses = helpers.sdk.send_and_check_request_objects(requests)
-
-    wallets = [helpers.wallet.create_client_wallet(seed) for seed in seeds]
-
-    yield trustee_wallets + wallets
-
-    # TODO: Not certain if this is actually changing the role.
-    def _update_nym_standard_user(response):
-        data = get_payload_data(response[RESULT])
-        request = helpers.request.nym(
-            dest=data[TARGET_NYM],
-            verkey=data[VERKEY],
-            role=None
-        )
-        return request
-
-    requests = [
-        _update_nym_standard_user(response)
-        for _, response in responses
-    ]
-
-    helpers.sdk.send_and_check_request_objects(requests)
